requirements.txt
pytest
aiohttp
pymodbus
pytest-asyncio
homeassistant
pytest-homeassistant-custom-componentpymodbus>=3.6.3
